Replace long polling way of getting updates with webhook
- Extend webhook class to deploy http server with `ngrok` redirect
