+++ src/main/java/Main.java
@@ -1,4 +1,4 @@
-import dev.aubique.jcalc.bot.PollingBot;
+import dev.aubique.jcalc.bot.BotHandler;
 import org.telegram.telegrambots.ApiContextInitializer;
 import org.telegram.telegrambots.meta.TelegramBotsApi;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiRequestException;
@@ -7,9 +7,10 @@ public class Main {
 
     public static void main(String[] args) {
         ApiContextInitializer.init();
-        TelegramBotsApi telegramBotApi = new TelegramBotsApi();
+        TelegramBotsApi telegramBotApi = null;
         try {
-            telegramBotApi.registerBot(PollingBot.getInstance());
+            telegramBotApi = new TelegramBotsApi("https://6537d7ed.eu.ngrok.io", "https://localhost:8443");
+            telegramBotApi.registerBot(BotHandler.getInstance());
         } catch (TelegramApiRequestException e) {
             e.printStackTrace();
         }
